refactor(webapp): extract tray-leader broadcast module
