+++ packages/webapp/src/scoops/tray-leader-sync.ts
@@ -5,13 +5,12 @@
 
 import type {
   LeaderToWorkerControlMessage,
-  TranscriptExportErrorCode,
   TranscriptExportSelector,
   WorkerBridgeCdpResponse,
   WorkerBridgeConnected,
   WorkerBridgeDisconnected,
 } from '@slicc/shared-ts';
-import { base64ToUint8, TranscriptExportError, uint8ToBase64 } from '@slicc/shared-ts';
+import { base64ToUint8, uint8ToBase64 } from '@slicc/shared-ts';
 import type { BrowserAPI } from '../cdp/browser-api.js';
 import { PreviewBridgeCdpTransport } from '../cdp/preview-bridge-cdp-transport.js';
 import { type RemoteCDPSender, RemoteCDPTransport } from '../cdp/remote-cdp-transport.js';
@@ -33,6 +32,7 @@ import {
   FollowerRegistry,
   labelForFollower,
 } from './tray-leader/follower-registry.js';
+import { TranscriptExportManager } from './tray-leader/transcript-export.js';
 import {
   CHERRY_RUNTIME_TAG,
   type CherryHostEventMessage,
@@ -169,15 +169,8 @@ export interface LeaderSyncManagerOptions {
     estimatedBytes?: number;
   }) => Promise<boolean>;
   /**
-   * True when this leader tab has no interactive human of its own — the
-   * hosted-leader (cloud) float, where the leader is headless Chromium inside
-   * an e2b sandbox. Such a leader cannot render `requestTranscriptExportApproval`
-   * to anyone, so it delegates the prompt to the requesting follower (whose
-   * user is the only human in the session) over the tray data channel.
-   *
-   * The gate is delegated, never skipped: the same dialog is shown, just in the
-   * realm where a human can actually answer it. Every failure mode — no reply,
-   * disconnect, unwired handler, dialog error — resolves to deny.
+   * True when this leader tab has no interactive human of its own. The approval
+   * gate is delegated to the requesting follower rather than skipped.
    */
   headlessLeader?: boolean;
   /**
@@ -236,65 +229,6 @@ export function selectTeleportPool<
   });
 }
 
-// ---------------------------------------------------------------------------
-// Transcript export helpers (module-level, pure functions)
-// ---------------------------------------------------------------------------
-
-interface SendExportChunksCtx {
-  bootstrapId: string;
-  requestId: string;
-  chunks: AsyncIterable<Uint8Array>;
-  hasher: { update(data: Uint8Array): void };
-  abort: AbortController;
-  onSlice: (index: number, slice: string) => boolean;
-  /** When true, leader waits for follower ack before sending the next slice. */
-  awaitAck: boolean;
-}
-
-interface SendExportChunksResult {
-  streamError: boolean;
-  chunkCount: number;
-  byteLength: number;
-}
-
-/**
- * Encode a Uint8Array to base64. Uses the browser btoa path for small arrays
- * and a chunked approach for large ones to avoid stack overflows from large
- * spread arguments to String.fromCharCode.
- */
-function bytesToBase64(data: Uint8Array): string {
-  // 8 KiB is a safety ceiling for JS engine call-stack argument limits, not a
-  // performance tuning knob. Do not confuse this with the 32 KiB wire chunk
-  // limit (EXPORT_CHUNK_B64_MAX): lowering this value only increases loop
-  // iterations, not throughput.
-  const CHUNK = 8192;
-  let binary = '';
-  for (let i = 0; i < data.length; i += CHUNK) {
-    binary += String.fromCharCode(...data.subarray(i, i + CHUNK));
-  }
-  return btoa(binary);
-}
-
-/**
- * Wait until the channel's bufferedAmount drops below `threshold`.
- * Uses 25 ms polling regardless of channel type. The real RTCDataChannel
- * `bufferedamountlow` event would be ~25× more responsive, but adding
- * cancellable listener cleanup requires `removeEventListener` on
- * `TrayDataChannelLike` — deferred as future work. For test doubles that
- * expose `bufferedAmount = 0`, the while-loop body never executes. For test
- * doubles with no `bufferedAmount` property, returns immediately.
- */
-async function waitForBufferedAmountLow(
-  channel: { bufferedAmount?: number },
-  threshold: number,
-  signal: AbortSignal
-): Promise<void> {
-  if (typeof channel.bufferedAmount !== 'number') return;
-  while (!signal.aborted && channel.bufferedAmount > threshold) {
-    await new Promise<void>((res) => setTimeout(res, 25));
-  }
-}
-
 /** Tracks a leader-initiated remote exec (the `ssh` command) awaiting the follower's streamed reply. */
 interface PendingRemoteExec {
   bootstrapId: string;
@@ -350,6 +284,7 @@ export class LeaderSyncManager {
   private readonly followerRegistry: FollowerRegistry;
   private readonly context: LeaderSyncContext;
   private readonly broadcast: BroadcastManager;
+  private readonly transcriptExport: TranscriptExportManager;
   private readonly registry = new TrayTargetRegistry();
   private get followers(): Map<string, ConnectedFollower> {
     return this.context.followers.followers;
@@ -417,93 +352,6 @@ export class LeaderSyncManager {
   private readonly previewLickLastEmitAt = new Map<string, number>();
   private static readonly PREVIEW_LICK_THROTTLE_MS = 2000;
 
-  /**
-   * In-flight transcript export requests keyed by composite `${bootstrapId}:${requestId}`.
-   * The composite key prevents cross-follower collisions when two followers happen to
-   * choose the same requestId. Wire messages still use requestId only; `exportKey()` builds
-   * the composite key from the bootstrapId known from the established connection.
-   * Cleared on deny/cancel/complete/error/disconnect.
-   */
-  private readonly activeExports = new Map<
-    string,
-    { bootstrapId: string; abort: AbortController }
-  >();
-
-  /**
-   * Ack waiters for the bounded in-flight chunk window (Wave 4).
-   *
-   * Keyed by `${bootstrapId}:${requestId}:${index}`. Each entry holds a
-   * resolver that is called when the matching `transcript.export.ack`
-   * arrives from the correct follower. Entries are rejected on abort or
-   * disconnect and deleted on receipt or cleanup.
-   */
-  private readonly ackWaiters = new Map<
-    string,
-    { resolve: () => void; reject: (err: Error) => void }
-  >();
-
-  /**
-   * Pending delegated approvals, keyed by the same composite
-   * `${bootstrapId}:${requestId}` as {@link activeExports}. Settled by the
-   * follower's `transcript.export.approve.response`, by the timeout, or by
-   * disconnect cleanup in {@link removeFollower}.
-   */
-  private readonly approvalWaiters = new Map<string, (approved: boolean) => void>();
-
-  /** Build the composite map key from follower identity and wire request ID. */
-  private static exportKey(bootstrapId: string, requestId: string): string {
-    return `${bootstrapId}:${requestId}`;
-  }
-
-  /** Build the ack waiter map key (owner-scoped). */
-  private static ackKey(bootstrapId: string, requestId: string, index: number): string {
-    return `${bootstrapId}:${requestId}:${index}`;
-  }
-
-  /**
-   * Reject and remove all ack waiters for a given bootstrapId:requestId.
-   * Called on cancel, disconnect, and abort to prevent stuck Promises.
-   */
-  private clearAckWaiters(bootstrapId: string, requestId: string): void {
-    const prefix = `${bootstrapId}:${requestId}:`;
-    for (const [key, waiter] of this.ackWaiters) {
-      if (key.startsWith(prefix)) {
-        waiter.reject(new Error('export aborted'));
-        this.ackWaiters.delete(key);
-      }
-    }
-  }
-
-  /** Max payload per chunk message: 32 KiB of base64 text. */
-  private static readonly EXPORT_CHUNK_B64_MAX = 32 * 1024;
-
-  /** Backpressure threshold: pause when bufferedAmount exceeds 1 MiB. */
-  private static readonly EXPORT_BACKPRESSURE_THRESHOLD = 1024 * 1024;
-
-  /**
-   * Per-ack deadline: if a follower's OPFS write stalls for longer than this
-   * (I/O contention, quota pressure, frozen tab) and the data channel stays
-   * alive, the export is aborted so the leader never waits indefinitely.
-   */
-  private static readonly ACK_TIMEOUT_MS = 30_000;
-
-  /**
-   * Deadline for a delegated approval reply. Generous because a human has to
-   * read the dialog and click, but bounded so a follower that never answers
-   * (closed tab, dead UI) cannot pin the export slot forever — the leader
-   * always sends a terminal `denied`.
-   */
-  private static readonly APPROVAL_TIMEOUT_MS = 120_000;
-
-  /**
-   * Minimum peer protocol version that supports transcript.export.ack (Wave 4).
-   * Pin to 3 — do NOT replace with TRAY_SYNC_PROTOCOL_VERSION. When the
-   * protocol bumps to v4+, v3 peers still understand acks and must remain
-   * ack-gated. Only bump this constant if a future wave removes ack support
-   * for v3 peers.
-   */
-  private static readonly ACK_PROTOCOL_VERSION_MIN = 3;
-
   constructor(private readonly options: LeaderSyncManagerOptions) {
     this.followerRegistry = new FollowerRegistry({
       log,
@@ -538,17 +386,11 @@ export class LeaderSyncManager {
       },
       afterRegistryCleanup: (bootstrapId) => {
         if (this.registry.hasChanged()) this.broadcastTargetRegistry();
-        for (const [key, entry] of this.activeExports) {
-          if (entry.bootstrapId !== bootstrapId) continue;
-          entry.abort.abort();
-          const requestId = key.slice(bootstrapId.length + 1);
-          this.clearAckWaiters(bootstrapId, requestId);
-          this.activeExports.delete(key);
-        }
-        for (const key of [...this.approvalWaiters.keys()]) {
-          if (key.startsWith(`${bootstrapId}:`)) this.settleApproval(key, false);
-        }
       },
+    });
+    this.transcriptExport = new TranscriptExportManager(this.context);
+    Object.defineProperty(this, 'activeExports', {
+      get: () => this.transcriptExport.activeExports,
     });
   }
 
@@ -1169,16 +1011,24 @@ export class LeaderSyncManager {
         this.handleFollowerExecMessage(bootstrapId, message);
         break;
       case 'transcript.export.request':
-        void this.handleTranscriptExportRequest(bootstrapId, message.requestId, message.selector);
+        void this.transcriptExport.handleTranscriptExportRequest(
+          bootstrapId,
+          message.requestId,
+          message.selector
+        );
         break;
       case 'transcript.export.cancel':
-        this.handleTranscriptExportCancel(bootstrapId, message.requestId);
+        this.transcriptExport.handleTranscriptExportCancel(bootstrapId, message.requestId);
         break;
       case 'transcript.export.ack':
-        this.handleTranscriptExportAck(bootstrapId, message.requestId, message.index);
+        this.transcriptExport.handleTranscriptExportAck(
+          bootstrapId,
+          message.requestId,
+          message.index
+        );
         break;
       case 'transcript.export.approve.response':
-        this.handleTranscriptExportApprovalResponse(
+        this.transcriptExport.handleTranscriptExportApprovalResponse(
           bootstrapId,
           message.requestId,
           message.approved
@@ -1246,520 +1096,6 @@ export class LeaderSyncManager {
     } else {
       log.info('Follower hello', { bootstrapId, protocolVersion: message.protocolVersion });
     }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Transcript export (leader side)
-  // ---------------------------------------------------------------------------
-
-  private async handleTranscriptExportRequest(
-    bootstrapId: string,
-    requestId: string,
-    selector: TranscriptExportSelector
-  ): Promise<void> {
-    // Composite key: ${bootstrapId}:${requestId}. Prevents cross-follower collisions
-    // when two different followers happen to choose the same requestId.
-    const key = LeaderSyncManager.exportKey(bootstrapId, requestId);
-
-    // One-use guard: same follower+requestId pair is a duplicate.
-    if (this.activeExports.has(key)) {
-      log.warn('Duplicate transcript export request from same follower, ignoring', {
-        bootstrapId,
-        requestId,
-      });
-      return;
-    }
-
-    const follower = this.followers.get(bootstrapId);
-    if (!follower) return;
-
-    const { requestTranscriptExportApproval, createTranscriptExport, headlessLeader } =
-      this.options;
-    // A headless leader has no dialog to show, so it only needs the ZIP factory
-    // — the prompt is delegated to the follower.
-    if ((!requestTranscriptExportApproval && !headlessLeader) || !createTranscriptExport) {
-      follower.sync.send({ type: 'transcript.export.denied', requestId });
-      return;
-    }
-
-    // Per-follower concurrency cap: at most one approved or pending export per follower.
-    // Auto-deny additional requests without opening approval or leaking metadata.
-    const hasInFlight = [...this.activeExports.values()].some((e) => e.bootstrapId === bootstrapId);
-    if (hasInFlight) {
-      log.warn('Follower already has an in-flight export; denying duplicate', {
-        bootstrapId,
-        requestId,
-      });
-      follower.sync.send({ type: 'transcript.export.denied', requestId });
-      return;
-    }
-
-    // Register in-flight state with AbortController before the async approval
-    const abort = new AbortController();
-    this.activeExports.set(key, { bootstrapId, abort });
-
-    // Derive follower identity from connected state (never trust request payload)
-    const followerLabel = labelForFollower(follower.floatType, follower.runtime);
-
-    // Send pending immediately to unblock the follower's UI
-    follower.sync.send({ type: 'transcript.export.pending', requestId });
-
-    let approved = false;
-    try {
-      if (headlessLeader) {
-        // No human in this tab — ask the follower's user instead.
-        approved = await this.requestDelegatedApproval(bootstrapId, requestId, selector);
-      } else if (requestTranscriptExportApproval) {
-        approved = await requestTranscriptExportApproval({
-          requestId,
-          followerLabel,
-          // Derived from connected state via targets.advertise — never from request
-          // payload. Only populated for Cherry followers; omitted otherwise.
-          hostOrigin: follower.hostOrigin,
-          selector,
-        });
-      }
-    } catch (err) {
-      log.warn('transcript export approval threw', {
-        requestId,
-        error: err instanceof Error ? err.message : String(err),
-      });
-      approved = false;
-    }
-
-    // Re-fetch follower — might have disconnected during approval dialog
-    const stillConnected = this.followers.get(bootstrapId);
-    if (!this.activeExports.has(key) || !stillConnected) {
-      // Cancelled or disconnected during approval
-      this.activeExports.delete(key);
-      return;
-    }
-
-    if (!approved) {
-      // Send denied — NO metadata about the transcript
-      stillConnected.sync.send({ type: 'transcript.export.denied', requestId });
-      this.activeExports.delete(key);
-      return;
-    }
-
-    // Approved: create export and stream chunks
-    void this.streamTranscriptExport(bootstrapId, requestId, selector, abort);
-  }
-
-  private handleTranscriptExportCancel(bootstrapId: string, requestId: string): void {
-    const key = LeaderSyncManager.exportKey(bootstrapId, requestId);
-    const entry = this.activeExports.get(key);
-    // Owner-aware guard: composite key already encodes bootstrapId, so only the
-    // originating follower's key exists. Guard against a missing entry anyway.
-    if (!entry) return;
-    log.info('Transcript export cancelled by follower', { requestId, bootstrapId });
-    entry.abort.abort();
-    this.activeExports.delete(key);
-    // Reject any ack waiter for this request so the sendExportChunks loop exits.
-    this.clearAckWaiters(bootstrapId, requestId);
-    // Settle a pending delegated approval so it does not idle until its
-    // timeout after the request it belongs to is already gone.
-    this.settleApproval(key, false);
-  }
-
-  /**
-   * Ask the requesting follower to render the approval dialog and resolve with
-   * the human's verdict. Used only by a headless (hosted / cloud) leader.
-   *
-   * Fail-closed: resolves `false` if the follower is gone, never replies within
-   * {@link APPROVAL_TIMEOUT_MS}, or the request is cancelled meanwhile.
-   */
-  private requestDelegatedApproval(
-    bootstrapId: string,
-    requestId: string,
-    selector: TranscriptExportSelector
-  ): Promise<boolean> {
-    const follower = this.followers.get(bootstrapId);
-    if (!follower) return Promise.resolve(false);
-
-    const key = LeaderSyncManager.exportKey(bootstrapId, requestId);
-    return new Promise<boolean>((resolve) => {
-      const timer = setTimeout(() => {
-        log.warn('Delegated transcript export approval timed out', { bootstrapId, requestId });
-        this.settleApproval(key, false);
-      }, LeaderSyncManager.APPROVAL_TIMEOUT_MS);
-
-      this.approvalWaiters.set(key, (approved) => {
-        clearTimeout(timer);
-        resolve(approved);
-      });
-
-      // `send` reports failure by returning false (it never throws). A prompt
-      // that never left the leader can never be answered — deny now rather than
-      // holding the export slot until the deadline.
-      const sent = follower.sync.send({
-        type: 'transcript.export.approve.request',
-        requestId,
-        selector,
-      });
-      if (!sent) {
-        log.warn('Failed to send delegated approval prompt — denying', {
-          bootstrapId,
-          requestId,
-        });
-        this.settleApproval(key, false);
-      }
-    });
-  }
-
-  /**
-   * Resolve a pending delegated approval, if one is still registered for `key`.
-   * Idempotent — a late or duplicate reply after the waiter is gone is a no-op.
-   */
-  private settleApproval(key: string, approved: boolean): void {
-    const waiter = this.approvalWaiters.get(key);
-    if (!waiter) return;
-    this.approvalWaiters.delete(key);
-    waiter(approved);
-  }
-
-  /** Handle `transcript.export.approve.response` from a follower. */
-  private handleTranscriptExportApprovalResponse(
-    bootstrapId: string,
-    requestId: string,
-    approved: boolean
-  ): void {
-    // Composite key scopes the verdict to the follower that was asked, so one
-    // follower can never answer another's prompt.
-    this.settleApproval(LeaderSyncManager.exportKey(bootstrapId, requestId), approved === true);
-  }
-
-  /**
-   * Handle `transcript.export.ack` from a follower.
-   * Resolves the waiter for the matching bootstrapId:requestId:index.
-   * Acks from other followers are silently dropped (owner-scoped key).
-   */
-  private handleTranscriptExportAck(bootstrapId: string, requestId: string, index: number): void {
-    const key = LeaderSyncManager.ackKey(bootstrapId, requestId, index);
-    const waiter = this.ackWaiters.get(key);
-    if (!waiter) return; // late or spurious ack — ignore
-    this.ackWaiters.delete(key);
-    waiter.resolve();
-  }
-
-  /**
-   * Stream a ZIP result over the data channel with bounded backpressure.
-   * Each raw-bytes slice is base64-encoded and split so each message carries
-   * at most EXPORT_CHUNK_B64_MAX bytes of base64 text.
-   */
-  private async streamTranscriptExport(
-    bootstrapId: string,
-    requestId: string,
-    selector: TranscriptExportSelector,
-    abort: AbortController
-  ): Promise<void> {
-    const key = LeaderSyncManager.exportKey(bootstrapId, requestId);
-    const { createTranscriptExport } = this.options;
-    const follower = this.followers.get(bootstrapId);
-    if (!follower || !createTranscriptExport) {
-      this.activeExports.delete(key);
-      return;
-    }
-    const sendErr = (code: TranscriptExportErrorCode): void => {
-      const f = this.followers.get(bootstrapId);
-      if (f) f.sync.send({ type: 'transcript.export.error', requestId, code });
-      this.activeExports.delete(key);
-    };
-
-    let result: TranscriptZipResult;
-    try {
-      result = await createTranscriptExport(selector, abort.signal);
-    } catch (createErr) {
-      if (abort.signal.aborted) {
-        this.activeExports.delete(key);
-        return;
-      }
-      log.warn('createTranscriptExport failed', {
-        requestId,
-        error: createErr instanceof Error ? createErr.message : String(createErr),
-      });
-      const code: TranscriptExportErrorCode =
-        createErr instanceof TranscriptExportError ? createErr.code : 'session-not-found';
-      sendErr(code);
-      return;
-    }
-
-    if (abort.signal.aborted || !this.activeExports.has(key)) {
-      this.activeExports.delete(key);
-      return;
-    }
-    const sync = this.followers.get(bootstrapId)?.sync;
-    if (!sync) {
-      this.activeExports.delete(key);
-      return;
-    }
-
-    sync.send({ type: 'transcript.export.start', requestId, filename: result.filename });
-
-    const { sha256: sha256Lib } = await import('js-sha256');
-    const hasher = sha256Lib.create();
-
-    // Gate ack-based flow control on the follower's protocol version.
-    // Followers that sent `hello` with protocolVersion >= 3 support acks;
-    // older or legacy peers (no hello) skip ack waiting for compatibility.
-    const followerForAck = this.followers.get(bootstrapId);
-    const awaitAck =
-      (followerForAck?.peerProtocolVersion ?? 0) >= LeaderSyncManager.ACK_PROTOCOL_VERSION_MIN;
-
-    const {
-      streamError,
-      chunkCount: chunkIndex,
-      byteLength: leaderByteCount,
-    } = await this.sendExportChunks({
-      bootstrapId,
-      requestId,
-      chunks: result.chunks,
-      hasher,
-      abort,
-      awaitAck,
-      onSlice: (idx, slice) => {
-        const f = this.followers.get(bootstrapId);
-        if (!f) return false;
-        f.sync.send({ type: 'transcript.export.chunk', requestId, index: idx, data: slice });
-        return true;
-      },
-    });
-    if (streamError) {
-      // Stream iteration failed — not a user abort.
-      sendErr('transfer-corrupt');
-      return;
-    }
-
-    // If the key is gone the export was cancelled or the follower disconnected
-    // (both paths delete the key before aborting) — return silently.
-    if (!this.activeExports.has(key)) return;
-    if (abort.signal.aborted) {
-      // Key still present ⇒ ack-timeout abort: notify the follower so it can
-      // clean up its in-flight OPFS spool.
-      sendErr('transfer-aborted');
-      return;
-    }
-
-    let completion: { byteLength: number; sha256: string };
-    try {
-      completion = await result.completion;
-    } catch {
-      if (abort.signal.aborted) {
-        this.activeExports.delete(key);
-        return;
-      }
-      sendErr('transfer-aborted');
-      return;
-    }
-
-    if (abort.signal.aborted || !this.activeExports.has(key)) {
-      this.activeExports.delete(key);
-      return;
-    }
-
-    // Verify the service's reported byte count matches what the leader actually sent.
-    // Both should equal the raw ZIP byte count; a mismatch indicates a service bug
-    // or a corrupted stream before the follower ever validates.
-    if (completion.byteLength !== leaderByteCount) {
-      log.warn('Transcript export byte count mismatch between service and leader stream', {
-        requestId,
-        serviceByteLength: completion.byteLength,
-        leaderByteCount,
-      });
-      sendErr('transfer-corrupt');
-      return;
-    }
-
-    // Cross-check leader-computed SHA-256 against the service's reported digest.
-    // Both cover the same raw ZIP bytes; a mismatch indicates a service/stream bug.
-    const leaderSha = hasher.hex();
-    if (completion.sha256 !== leaderSha) {
-      log.warn('Transcript export SHA-256 mismatch between service and leader stream', {
-        requestId,
-      });
-      sendErr('transfer-corrupt');
-      return;
-    }
-
-    const done = this.followers.get(bootstrapId);
-    if (done) {
-      done.sync.send({
-        type: 'transcript.export.complete',
-        requestId,
-        chunks: chunkIndex,
-        byteLength: leaderByteCount,
-        sha256: leaderSha,
-      });
-    }
-    this.activeExports.delete(key);
-  }
-
-  /**
-   * Iterate the ZIP chunk iterable, base64-encode each slice (≤32 KiB), apply
-   * backpressure, and call `ctx.onSlice` per slice.
-   *
-   * Returns a result object: `streamError` is true when a stream error occurred
-   * (caller should send `transfer-corrupt`); false when completed normally OR
-   * aborted (caller distinguishes via `abort.signal.aborted` or
-   * `activeExports.has(key)`). `chunkCount` and `byteLength` reflect the
-   * totals processed regardless of terminal state.
-   */
-  private async sendExportChunks(ctx: SendExportChunksCtx): Promise<SendExportChunksResult> {
-    const { bootstrapId, requestId, chunks, hasher, abort, onSlice, awaitAck } = ctx;
-    const key = LeaderSyncManager.exportKey(bootstrapId, requestId);
-    const sync = this.followers.get(bootstrapId)?.sync;
-    if (!sync) return { streamError: true, chunkCount: 0, byteLength: 0 };
-    let idx = 0;
-    let byteCount = 0;
-    try {
-      for await (const raw of chunks) {
-        if (abort.signal.aborted || !this.activeExports.has(key)) {
-          return { streamError: false, chunkCount: idx, byteLength: byteCount };
-        }
-        hasher.update(raw);
-        byteCount += raw.byteLength;
-        const b64 = bytesToBase64(raw);
-        const sliceResult = await this.sendExportSlices({
-          bootstrapId,
-          requestId,
-          b64,
-          sync,
-          key,
-          abort,
-          onSlice,
-          awaitAck,
-          startIdx: idx,
-        });
-        idx = sliceResult.nextIdx;
-        if (sliceResult.done !== 'continue') {
-          return {
-            streamError: sliceResult.done === 'error',
-            chunkCount: idx,
-            byteLength: byteCount,
-          };
-        }
-      }
-    } catch (streamErr) {
-      if (abort.signal.aborted) {
-        return { streamError: false, chunkCount: idx, byteLength: byteCount };
-      }
-      log.warn('Error streaming transcript export chunks', {
-        requestId,
-        error: streamErr instanceof Error ? streamErr.message : String(streamErr),
-      });
-      return { streamError: true, chunkCount: idx, byteLength: byteCount };
-    }
-    return { streamError: false, chunkCount: idx, byteLength: byteCount };
-  }
-
-  /** Send all base64 slices for one raw ZIP chunk, applying backpressure and optional ack-gating. */
-  private async sendExportSlices(ctx: {
-    bootstrapId: string;
-    requestId: string;
-    b64: string;
-    sync: { bufferedAmount?: number };
-    key: string;
-    abort: AbortController;
-    onSlice: (index: number, slice: string) => boolean;
-    awaitAck: boolean;
-    startIdx: number;
-  }): Promise<{ nextIdx: number; done: 'continue' | 'abort' | 'error' }> {
-    const { bootstrapId, requestId, b64, sync, key, abort, onSlice, awaitAck } = ctx;
-    let idx = ctx.startIdx;
-    for (let off = 0; off < b64.length; off += LeaderSyncManager.EXPORT_CHUNK_B64_MAX) {
-      if (abort.signal.aborted || !this.activeExports.has(key)) {
-        return { nextIdx: idx, done: 'abort' };
-      }
-      await waitForBufferedAmountLow(
-        sync,
-        LeaderSyncManager.EXPORT_BACKPRESSURE_THRESHOLD,
-        abort.signal
-      );
-      if (abort.signal.aborted || !this.activeExports.has(key)) {
-        return { nextIdx: idx, done: 'abort' };
-      }
-      const slice = b64.slice(off, off + LeaderSyncManager.EXPORT_CHUNK_B64_MAX);
-      if (!onSlice(idx, slice)) {
-        return { nextIdx: idx, done: 'error' };
-      }
-      // Wave 4: wait for durable-write ack from v3+ followers before sending the next slice.
-      if (awaitAck) {
-        const waited = await this.waitForAck(bootstrapId, requestId, idx, abort);
-        if (!waited || abort.signal.aborted || !this.activeExports.has(key)) {
-          return { nextIdx: idx + 1, done: 'abort' };
-        }
-      }
-      idx++;
-    }
-    return { nextIdx: idx, done: 'continue' };
-  }
-
-  /**
-   * Register an ack waiter for the given chunk index and await it.
-   * Returns `true` when ack arrives, `false` when aborted/cancelled/timed-out.
-   * The waiter is stored with an owner-scoped key so cross-follower
-   * acks are silently dropped by `handleTranscriptExportAck`.
-   *
-   * A 30 s deadline timer is registered so a follower whose OPFS write stalls
-   * indefinitely (live channel, stuck I/O) does not hold the export forever.
-   * When the timer fires, the export's AbortController is triggered and all
-   * sibling ack waiters for this export are rejected via clearAckWaiters.
-   */
-  private waitForAck(
-    bootstrapId: string,
-    requestId: string,
-    index: number,
-    abort: AbortController
-  ): Promise<boolean> {
-    const signal = abort.signal;
-    return new Promise<boolean>((resolve) => {
-      if (signal.aborted) {
-        resolve(false);
-        return;
-      }
-
-      const key = LeaderSyncManager.ackKey(bootstrapId, requestId, index);
-      let timer: ReturnType<typeof setTimeout> | undefined;
-
-      const clearTimer = (): void => {
-        if (timer !== undefined) {
-          clearTimeout(timer);
-          timer = undefined;
-        }
-      };
-      const cleanup = (): void => {
-        this.ackWaiters.delete(key);
-        clearTimer();
-      };
-      const onAbort = (): void => {
-        cleanup();
-        resolve(false);
-      };
-      signal.addEventListener('abort', onAbort, { once: true });
-
-      // Per-ack deadline: abort the export if the follower stalls.
-      // abort.abort() fires the signal’s abort event synchronously, which
-      // triggers onAbort (cleanup + resolve false). clearAckWaiters then
-      // handles any sibling waiters for this export.
-      timer = setTimeout(() => {
-        timer = undefined;
-        abort.abort();
-        this.clearAckWaiters(bootstrapId, requestId);
-      }, LeaderSyncManager.ACK_TIMEOUT_MS);
-
-      this.ackWaiters.set(key, {
-        resolve: () => {
-          signal.removeEventListener('abort', onAbort);
-          cleanup();
-          resolve(true);
-        },
-        reject: () => {
-          signal.removeEventListener('abort', onAbort);
-          cleanup();
-          resolve(false);
-        },
-      });
-    });
   }
 
   /**
